Extract discovery package
