+++ common.py
@@ -1,0 +1,24 @@
+"""Common elements used almost all over the code
+
+Copyright 2018 Spectre Team
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+import os
+from typing import Tuple
+
+FILESYSTEM_ROOT = os.path.abspath(os.sep)
+DATASETS_ROOT = os.path.join(FILESYSTEM_ROOT, 'data')
+
+Response = Tuple[str, int]
+NOT_FOUND = "", 404
